==> 02.integration_mc_rna/02.process_AIBS_TENX.ipynb <==
import numpy as np
import pandas as pd

import anndata
import scanpy as sc
from ALLCools.mcds import MCDS
from wmb import aibs


aibs_annot = aibs.get_tenx_annot()
aibs_annot

annot = aibs_annot[['L1', 'L2', 'L3', 'L1_annot', 'L2_annot', 'DissectionRegion', 'SubRegion', 'MajorRegion']].to_pandas()

NN_types = ['NN', 'IMN']
nonoverlap_types = np.concatenate([annot['L2_annot'][annot['L2_annot'].str.startswith('CB')].unique(),
                                   annot['L2_annot'][annot['L2_annot'].str.startswith('DCO')].unique(),
                                   annot['L2_annot'][annot['L2_annot'].str.startswith('MM')].unique(),
                                  ])


rna_region_dict = {
    'CTX': ["RSP", "ACA","PL-ILA-ORB", "AI-CLA", "AId-AIv", "AId-AIv-AIp", "TEa-PERI-ECT", "AUD-TEa-PERI-ECT", "AUD", "PTLp", "VIS", "VIS-PTLp", "VISl", 'VISm', 'VISp', "MOp", "MOs-FRP", "MO-FRP", "SSp", "SSs-GU-VISC-AIp", "SS-GU-VISC"],
    'HIP': ["HIP", "HIP - CA"],
    'OLF': ["OLF - MOB-AOB", "OLF - AON-TT-DP"],
    'RHP': ["PAR-POST-PRE-SUB-ProS", 'ENT'],
    'PIR': ["OLF - COA-PAA-NLOT-TR","OLF - PIR", "OLF - AON-TT-DP-PIR-COA-NLOT-PAA-TR", 'CTXsp - CLA-EP-LA-BLA-BMA-PA'],
    'AMY': ["STR - sAMY"],
    'STR': ["STR - STRd", "STR - STRv"],
    'PAL': ["CNU - PAL"],
    'TH': ["TH - Middle-lateral", "TH - LGd-IGL-LGv", "TH - Posterior", "TH - Anterior", "TH - PVT-PT", "TH - MG-SPFp-PP-POL-SGN-PoT-PIL", "TH - PO-Eth", "TH - PF-SPA-SPFm-VPMpc-VPLpc-RT", "TH", "TH - MH-LH-LP","TH - Middle-medial", "TH - AD-AV-AM-IAD-LD", "TH - RE-RH-CM-IAM-SMT-PR-Xi", "TH - MD-IMD-PCN-CL", "TH - VAL-VPM-VPL-VM"], 
    'HY': ["HY - HYa1", "HY - HYa2", "HY - HYm2", "HY - HYml", "HY - HYmm", "HY - HYpl", "HY - HYpm", "HY - MEZ-PVZ-PVR", "HY LZ"], 
    'MB': ["MB - VTA-SN", "MB - IC", "MB - MRN-CUN-RN-RR-PPN-PRT-SC-IC","MB - SC-IC",'MB - SC',"MB - MRN-CUN-RN-RR-PN", "MB - MRN-CUN-RN-RR-PPN-NB-SAG-PBG", "MB - MRN-CUN-RN-RR-PPN-PRT", "MB - PAG-RAmb", 'MB - PRT'], 
    'HB': ["MY - AP-DCO-VCO-DCN-ECU-NTS", "MY - CN-DCN-ECU-NTS-SPVC-SPVI-SPVO", "MY - MYsat-MYmot", "MY - MYsat-MYmot ant", "MY - MYsat-MYmot post", "MY - MYsat-MYmot-vent post", "MY - MYsat-Mymot-vent ant", "MY - SPVC-SPVI-SPVO", "MY - VNC-PHY-DMX-XII","HB - MY", "PONS - Pmot-Psat ant", "PONS - Pmot-Psat post", "PONS - Psen"]
}


# Parameters
group_name = 'STR'


t = group_name
selregion = rna_region_dict[t]


# select all CB cells
selc = annot.loc[annot['DissectionRegion'].isin(selregion) & (~annot['L1_annot'].isin(NN_types)) & (~annot['L2_annot'].isin(nonoverlap_types))].index
len(selc)

rnads_path = aibs.AIBS_TENX_V2_ZARR_PATH
rna_ds = MCDS.open(rnads_path, use_obs=selc, var_dim='gene')
rna_ds

cef = pd.read_csv(f'{t}_mC_CEF.txt', header=None, index_col=0).index
cef.name = 'gene'
cef

gene_id_not_nan = ~rna_ds.get_index('gene').isna()
is_cef = rna_ds.get_index('gene').isin(cef)
use_genes = gene_id_not_nan & is_cef
use_genes.sum()

adata = rna_ds.get_count_adata(da_name='gene_da', use_vars=use_genes)
adata

# same as sc.pp.normalize_per_cell(adata) but this method only normalize to 
# total read counts of the remaining CEF 
# instead of all the gene counts as saved in umi_count

adata.X.data = adata.X.data/np.repeat(adata.obs['umi_count'].values, adata.X.getnnz(axis=1)
                                     ) * adata.obs['umi_count'].mean()


sc.pp.log1p(adata)
#sc.pp.scale(adata) this method is supposed to use large memory so skip this 

adata.var['mean'] = adata.X.mean(axis=0).A1
adata.var['std'] = (adata.X.multiply(adata.X)).mean(axis=0).A1 - (adata.var['mean'].values ** 2)
print(adata.var['std'].min())


adata.obs[annot.columns] = annot.loc[adata.obs.index,:]
adata

pass_std_filter = (adata.var['std'].values > 0.005)
adata._inplace_subset_var(pass_std_filter)


np.savetxt('use_genes.txt', adata.var.index, fmt='%s')


adata.write_h5ad(f'{t}_{adata.shape[0]}_rna.h5ad')
