==> 02.integration_mc_rna/03.process_RS2.ipynb <==
import pandas as pd

import anndata
import scanpy as sc
from ALLCools.mcds import MCDS
from ALLCools.clustering import log_scale


var_dim_er ='geneslop2k'
chrom_to_remove = ['chrX', 'chrY', 'chrM', 'chrL']
excluded_L1_annot = ['MGC', 'ODC', 'ASC', 'OPC', 'CB', 'CBX']
mc_type='CHN'


region_to_subregion = {'CTX': ['MOp','SSp','ACA','AI','RSP','AUD','PTLp','VIS'],
                       'HIP': ['CAa','CAp','DGa','DGp'],
                       'RHP': ['ENT'],
                       'OLF': ['MOB'],
                       'PIR': ['PIRa','PIRp'],
                       'STR': ['STR'],
                       'PAL': ['PAL'],
                       'AMY': ['AMY'],
                       'TH': ['THm','THl','THp'],
                       'HY': ['HY'],
                       'MB': ['SC','MRN','VTA','PAG','IC'],
                       'HB': ['P','MY']
                      }
rs2_subregion_dict = {
    'MOp': ['3C', '4B'],
    'SSp': ['6B', '7B'],
    'ACA': ['3A', '4A', '5A'],
    'AI': ['3D'],
    'RSP': ['9A', '10A'],
    'AUD': ['9D', '10C'],
    'PTLp': ['9B'],
    'VIS': ['11B', '12B'],
    'ENT': ['10D11D12D13B'],
    'CAa': ['8E9H'],
    'CAp': ['10E11E'],
    'DGa': ['8J9J'],
    'DGp': ['10F11F'],
    'PIRa': ['2D3E4F5G'],
    'PIRp': ['6D7D8D9E'],
    'MOB': ['1C', '2E'],
    'PAL': ['4H5H6F7F', '6F7F'],
    'STR': ['4D5E6E'],
    'AMY': ['7H8H9G'],
    'THl': ['7E8F'], 
    'THm': ['7G8G'], 
    'THp': ['9K10G'],
    'HY': ['6H7J8K9L'],
    'SC': ['11G12F13C'],
    'MRN': ['10H11H12H'],
    'VTA': ['10J11J'],
    'PAG': ['12G13D'],
    'IC': ['14A'],
    'P': ['12J13E13F14C14D'],
    'MY': ['15C16C17B18B']
}

group_name = 'STR'

t = group_name
selected_ER_slice = [x for xx in region_to_subregion[t] for x in rs2_subregion_dict[xx]]


meta = pd.read_csv('/cemba/CEMBA_RS2/cell_48032_RS2_meta_nooutlier.csv.gz', index_col=0, header=0)
meta


selc = meta.index[meta['Slice'].isin(selected_ER_slice) & ~meta['L1_annot'].isin(excluded_L1_annot)]
print(len(selc))


mcds = MCDS.open('/cemba/CEMBA_RS2/CEMBA.epiretro.mcds', var_dim=var_dim_er, use_obs=selc)
mcds


use_genes = pd.read_csv(f'use_genes.txt', header=None, index_col=0).index
use_genes

hvf_judge = mcds.get_index(mcds.var_dim)
hvf_judge = pd.Series(hvf_judge.isin(use_genes), index=hvf_judge)
hvf_judge

mcds.coords['_feature_select'] = hvf_judge
# mcds = mcds.sel({mcds.var_dim:use_genes})
er = mcds.get_adata(mc_type='CHN', select_hvf='_feature_select')
er

# log_scale use StandardScaler which normalized to unit variance 
log_scale(er, with_mean=True)
er.X = -er.X


er = er[:, use_genes]
er.obs = meta.loc[er.obs.index,:]


er.write_h5ad(f'{t}_{er.shape[0]}_rs2_mch.h5ad')
