==> 02.integration_mc_rna/05.integrate.ipynb <==
import time
import glob
import numpy as np
from sklearn.preprocessing import normalize

import anndata
from ALLCools.integration.seurat_class import SeuratIntegration


# Parameters
group_name = "STR"


t = group_name

npc = 50
ncc = 50

# Integrate RS1 mC and RNA


integrator = SeuratIntegration()

ref_adata = anndata.read_h5ad(glob.glob(f'{t}_*_rs1_mch.h5ad')[0])
qry_adata = anndata.read_h5ad(glob.glob(f'{t}_*_rna.h5ad')[0])

adata_list = [ref_adata, qry_adata]
for xx in adata_list:
    xx.obsm['X_pca'] = normalize(xx.obsm['cef_pca'][:, :npc], axis=1)

adata_list

anchor = integrator.find_anchor(adata_list,
                                k_local=None,
                                key_local='X_pca',
                                k_anchor=5,
                                key_anchor='X',
                                dim_red='cca',
                                max_cc_cells=50000,
                                k_score=30,
                                k_filter=None,
                                scale1=False,
                                scale2=True,
                                #scale =[False, True]
                                n_components=ncc,
                                n_features=200,
                                alignments=[[[0], [1]]])


start_time = time.time()
corrected = integrator.integrate(key_correct='X_pca',
                                 row_normalize=True,
                                 n_components=npc,
                                 k_weight=100,
                                 sd=1,
                                 alignments=[[[0], [1]]])

print(time.time() - start_time)


qry_adata.obsm['X_pca_corrected'] = normalize(corrected[1], axis=1)
qry_adata.write_h5ad(f'{t}_{qry_adata.shape[0]}_rna.h5ad')


integrator.save(f'{t}_integration_rs1_rna')


# Integrate RS1 mC and RS2 EpiRetro

integrator = SeuratIntegration()

ref_adata = anndata.read_h5ad(glob.glob(f'{t}_*_rs1_mch.h5ad')[0])
qry_adata = anndata.read_h5ad(glob.glob(f'{t}_*_rs2_mch.h5ad')[0])

adata_list = [ref_adata, qry_adata]
for xx in adata_list:
    xx.obsm['X_pca'] = normalize(xx.obsm['cef_pca'][:, :npc], axis=1)

adata_list

start_time = time.time()
anchor = integrator.find_anchor(adata_list,
                                k_local=None,
                                key_local='X_pca',
                                k_anchor=5,
                                key_anchor='X',
                                dim_red='rpca',
                                max_cc_cells=100000,
                                k_score=30,
                                scale1=False,
                                scale2=False,
                                n_components=npc,
                                n_features=200,
                                alignments=[[[0], [1]]])
print(time.time() - start_time)


start_time = time.time()
corrected = integrator.integrate(key_correct='X_pca',
                                 row_normalize=True,
                                 n_components=npc,
                                 k_weight=100,
                                 sd=1,
                                 alignments=[[[0], [1]]])

print(time.time() - start_time)


qry_adata.obsm['X_pca_corrected'] = normalize(corrected[1], axis=1)
qry_adata.write_h5ad(f'{t}_{qry_adata.shape[0]}_rs2_mch.h5ad')


integrator.save(f'{t}_integration_rs1_rs2')
